requirements.txt
cement==2.10.2
certifi==2017.7.27.1
chardet==3.0.4
fusepy==2.0.4
idna==2.6
inflection==0.3.1
Jinja2==2.9.4
MarkupSafe==1.0
peewee==2.8.5
pycryptodome==3.4.7
requests==2.18.4
semantic-version==2.6.0
six==1.10.0
urllib3==1.22
